backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.4.1
langchain==0.1.12
langchain-google-genai==0.0.11
chromadb==0.4.24
pandas==2.2.1
pydantic==2.6.3
